Create lessons.db, movies.sqlite3, and project.db
The lessons.db contains the notes from lesson 2 in the SQL series.  The lesson included queries to practice the keywords AS, DISTINCT, WHERE, LIKE, BETWEEN, AND, OR, ORDER BY, ASC, DESC, LIMIT, and CASE. The project.db file contains the practice queries, the instructions for recreating the database, and the solutions to the practice queries. The movies.sqlite3 is the database that was used for the practice queries. If the content is corrupt the instructions to create the database has been included in Project.db.