adds fake_apps.sqlite3, Lessons.db, and Project.db
fake_apps.sqlite is the database used for the exercises in Lessons.db and in Project.db. Lessons.db contains the notes from the SQL lesson and Project.db contains the code needed to recreate the database if needed and the questions that relate to the clauses learned in this lesson.