add FinalProject.sql and modified previous files
this commit adds in the FinalProject.sql which is mean to be the indepent study portion of the learning process. I have added one query for each clause and 10 practice queries which utilize the animalHospital.sqlite database made in lesson 4's project. The other edits made were to correct spelling errors and update the database accordingly.